+++ packages/collector/test/tracing/logger/bunyan/app-instana-receives-non-bunyan-logger.js
@@ -33,7 +33,7 @@ const express = require('express');
 const morgan = require('morgan');
 
 const app = express();
-const logPrefix = `Express / Bunyan App [Instana receives non-Bunyan logger] (${process.pid}):\t`;
+const logPrefix = `Bunyan App [Instana receives non-Bunyan logger] (${process.pid}):\t`;
 
 if (process.env.WITH_STDOUT) {
   app.use(morgan(`${logPrefix}:method :url :status`));